+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to **bUnit** will be documented in this file. The project ad
 ### Changed
 - Better error description from `MarkupMatches` when two sets of markup are different.
 
+### Fixed
+- `cut.FindComponent<xxx>()` doesnt return the component in cut. It now searches and finds the first child component.
+
 ### Removed
 - The generic collection assertion methods `ShouldAllBe<T>(this IEnumerable<T> collection, params Action<T, int>[] elementInspectors)` and `ShouldAllBe<T>(this IEnumerable<T> collection, params Action<T>[] elementInspectors)` have been removed from the library.